requirements.txt
# Web framework
fastapi
uvicorn
httpx
apscheduler

# ORM and PostgreSQL
sqlalchemy
psycopg2-binary

# Pydantic for datavalidation
pydantic
pydantic-settings

# Fuzzy matching
rapidfuzz
unicode

# BLOB storage
minio
azure-storage-blob
google-cloud-storage
google-auth

# Vector store and embeddings
pymilvus
numpy
sentence-transformers

# Graph database
neo4j

# LLM support: Meta Llama 4 Maverick via Hugging Face
transformers
torch

# Async DB
asyncpg

# Testing
pytest
pytest-cov
pytest-asyncio
pytest-mock

# Utilities
python-dotenv
tqdm
lxml
requests
kubernetes